moved tests and model
