+++ model/OpenAPIContent.ts
