+++ tests/openapi/utils.test.ts
@@ -1,13 +1,13 @@
 import { readFileSync } from 'fs';
 import { load as loadYaml } from 'js-yaml';
 import { OpenAPIV3 } from 'openapi-types';
-import { getOpenApisContent } from '../../openapi';
+import { getOpenApisContent } from '../../src/openapi';
 
 function getPetstoreDocApi(path: string): OpenAPIV3.Document {
     return loadYaml(readFileSync(path, 'utf8')) as OpenAPIV3.Document
 }
 test('Petstore should be validated', async () => {
-    const petstorePath = "src/__tests__/resources/petstore/petstore.yaml"
-    const paths = [petstorePath, "src/__tests__/resources/petstore/ignore-me.yaml"]
+    const petstorePath = "tests/resources/petstore/petstore.yaml"
+    const paths = [petstorePath, "tests/resources/petstore/ignore-me.yaml"]
     expect(getOpenApisContent(paths)).resolves.toStrictEqual([{ path: petstorePath, content: getPetstoreDocApi(petstorePath) }])
 });
