+++ tests/resources/petstore/ignore-me.yaml
