+++ tests/resources/petstore/petstore.yaml
